// src/app/pages/uretim/uretim-emri/core/models/update-uretim-emri-model.ts
export class UpdateUretimEmriModel {

    id?: any;
    stokAdi?: any;
    birimAdi?: any;
    stokId?: any;
    urunReceteId?: any;
    urunReceteAdi?: any;
    planlananTarih?: any;
    sorumlu?: any;
    miktar?: any;
    hourId?: any;
    uretimEmriBilesenler?: any[];
    isEmirleri?: any[];
}